+++ next-docs/components/components.ts
@@ -1,0 +1,174 @@
+const components = [
+  {
+    name: 'Accordion',
+    href: '/core/accordion',
+    text: "An accordion is a vertical stack of interactive headings used to toggle the display of further information; each item can be 'collapsed' with just a short label visible or 'expanded' to show the full content.",
+  },
+  {
+    name: 'Alert',
+    href: '/core/alert',
+    text: 'A way of informing the user of important changes in a prominent way.',
+  },
+  {
+    name: 'AuthCode',
+    href: '/core/authCode',
+    text: 'A one-time password (OTP) is an automatically generated numeric or alphanumeric string of characters that authenticates a user for a single transaction or login session.',
+  },
+  {
+    name: 'Avatar',
+    href: '/core/avatar',
+    text: 'The Avatar component is typically used to display images, icons, or initials representing people or other entities.',
+  },
+  {
+    name: 'BottomSheet',
+    href: '/core/bottomSheet',
+    text: 'The bottom sheet component is a modified dialog that slides from the bottom of the screen, common pattern in mobile apps.',
+  },
+  {
+    name: 'Breadcrumb',
+    href: '/core/breadcrumb',
+    text: 'A list of links showing the location of the current page in the navigational hierarchy.',
+  },
+  {
+    name: 'Button',
+    href: '/core/button',
+    text: 'Buttons allow users to take actions, and make choices, with a single tap.',
+  },
+  {
+    name: 'IconButton',
+    href: '/core/iconButton',
+    text: 'Buttons allow users to take actions, and make choices, with a single tap.',
+  },
+  {
+    name: 'Carousel',
+    href: '/core/carousel',
+    text: 'Is an effective way of displaying multiple images or content in a single space.',
+  },
+  {
+    name: 'Checkbox',
+    href: '/core/checkbox',
+    text: 'The checkbox is shown as a square box that is ticked (checked) when activated.',
+  },
+  {
+    name: 'Chip',
+    href: '/core/chip',
+    text: 'Chips help filter content, or trigger actions.',
+  },
+  {
+    name: 'Drawer',
+    href: '/core/drawer',
+    text: 'The Drawer component is a panel that slides out from the edge of the screen. It can be useful when you need users to complete a task or view some details without leaving the current page.',
+  },
+  {
+    name: 'Dropdown',
+    href: '/core/dropdown',
+    text: 'Dropdowns is a custom select component that allows users to make single or multiple selections.',
+  },
+  {
+    name: 'Form',
+    href: '/core/form',
+    text: 'Form component is a grouping of input controls that allow a user to submit information to a server.',
+  },
+  {
+    name: 'Icons',
+    href: '/core/icons',
+    text: 'Moon DS provides a set of commonly used interface icons you can use in your project.',
+  },
+  {
+    name: 'Input',
+    href: '/core/input',
+    text: 'Text input fields allow users to enter text and can be used to collect user feedback or enter information in data entry forms.',
+  },
+  {
+    name: 'InsetInput',
+    href: '/core/insetInput',
+    text: 'Text input fields allow users to enter text and can be used to collect user feedback or enter information in data entry forms.',
+  },
+  {
+    name: 'Group',
+    href: '/core/group',
+    text: 'Combine different types of inputs into groups to save vertical space on your designs and also simplify form filling.',
+  },
+  {
+    name: 'Loader',
+    href: '/core/loader',
+    text: 'Loaders provide a visual cue that an action is processing awaiting a course of change or a result.',
+  },
+  {
+    name: 'MenuItem',
+    href: '/core/menuItem',
+    text: 'Menu items are used in such vertical menues and containers as Popovers, Sidebars, Drawers, Dialogs etc.',
+  },
+  {
+    name: 'Modal',
+    href: '/core/modal',
+    text: 'A modal is an interface element that appears over other content. It requires an interaction from the user before they can return to whatever is underneath.',
+  },
+  {
+    name: 'Popover',
+    href: '/core/popover',
+    text: 'Popovers are perfect for floating panels with arbitrary content like navigation menus, mobile menus and flyout menus.',
+  },
+  {
+    name: 'CircularProgress',
+    href: '/core/circularProgress',
+    text: 'A progress indicator (Circular and Linear) is a visual representation of a user’s progress through a set of steps, guiding toward the completion of a specified process.',
+  },
+  {
+    name: 'Progress',
+    href: '/core/progress',
+    text: 'A progress indicator (Circular and Linear) is a visual representation of a user’s progress through a set of steps, guiding toward the completion of a specified process.',
+  },
+  {
+    name: 'Radio',
+    href: '/core/radio',
+    text: 'Radio buttons are used to represent a group of choices whereby users can only select one option.',
+  },
+  {
+    name: 'Search',
+    href: '/core/search',
+    text: 'Search enables users to specify a word or a phrase to find relevant pieces of content without the use of navigation.',
+  },
+  {
+    name: 'Select',
+    href: '/core/select',
+    text: 'A form input used for selecting a value: when collapsed it shows the currently selected option and when expanded, it shows a scrollable list of predefined options for the user to choose from.',
+  },
+  {
+    name: 'Snackbar',
+    href: '/core/snackbar',
+    text: 'The snackbar component is a non-disruptive message that appears on the interface to provide quick, at-a-glance feedback on the outcome of an action.',
+  },
+  {
+    name: 'Switch',
+    href: '/core/switch',
+    text: 'Switch is a control that is used to quickly switch between two possible states.',
+  },
+  {
+    name: 'Table',
+    href: '/core/table',
+    text: 'A component for displaying large amounts of data in rows and columns.',
+  },
+  {
+    name: 'Tabs',
+    href: '/core/tabs',
+    text: 'Tabs to allow users to navigate easily between views within the same context.',
+  },
+  {
+    name: 'Tag',
+    href: '/core/tag',
+    text: 'Tags represent a set of interactive keywords that help organise and categorise objects.',
+  },
+  {
+    name: 'Textarea',
+    href: '/core/textarea',
+    text: 'A form control for editing multi-line text.',
+  },
+  {
+    name: 'Tooltip',
+    href: '/core/tooltip',
+    text: 'A means of displaying a description or extra information about an element, usually on hover, but can also be on click or tap.',
+  },
+];
+
+export default components;