+++ next-docs/components/sidebar/RecursiveNavItem.tsx
@@ -38,6 +38,7 @@ const RecursiveNavItem: React.FC<Props> = ({ item, pathname, onClick }) => {
       </Accordion>
     );
   }
+  console.log('item', item);
   return (
     <Link key={name} href={href} isActive={isCurrent} onClick={onClick}>
       {name}