+++ next-docs/pages/core/circularProgress.tsx
@@ -1,5 +1,6 @@
 import Preview from '../../components/codePreview/Preview';
 import ComponentPageDescription from '../../components/ComponentPageDescription';
+import getComponent from '../../components/getComponent';
 import PropsTable from '../../components/PropsTable';
 import Colors from '../../public/examples/circularProgress/Colors';
 import Default from '../../public/examples/circularProgress/Default';
@@ -9,19 +10,17 @@ import useExamples from '../../utils/useExamples';
 
 const Example = () => {
   const examples = useExamples('circularProgress');
+  const { name, text, image } = getComponent('CircularProgress');
   return (
     <>
       <ComponentPageDescription
-        title="CircularProgress"
-        isAriaSupport
+        title={name}
+        image={image}
+        isInProgress
         isRtlSupport
-        isInProgress // Added 01.11.2022
+        isAriaSupport
       >
-        <p>
-          A circular progress indicator is a visual representation of a user's
-          progress through a set of steps, guiding toward the completion of a
-          specified process.
-        </p>
+        <p>{text}</p>
       </ComponentPageDescription>
       <Preview
         title="Default"