+++ packages/charts/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 9.26.3
+
+### Patch Changes
+
+- Update docs
+- Updated dependencies [undefined]
+  - @heathmont/moon-assets@9.26.3
+  - @heathmont/moon-components@9.26.3
+  - @heathmont/moon-core@9.26.3
+  - @heathmont/moon-icons@9.26.3
+  - @heathmont/moon-themes@9.26.3
+  - @heathmont/moon-utils@9.26.3
+
 ## 9.26.2
 
 ### Patch Changes