+++ workspaces/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @heathmont/moon-core-tw
 
+## 9.26.3
+
+### Patch Changes
+
+- Update docs
+
 ## 9.26.2
 
 ### Patch Changes